backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
Pillow==10.1.0
validators==0.22.0
dnspython==2.4.2
fastapi==0.95.2
uvicorn[standard]==0.22.0
# Optional: PyTorch is optional and can be installed per environment. For CPU-only:
# pip install "torch" --index-url https://download.pytorch.org/whl/cpu
# For GPU/CUDA install follow official instructions: https://pytorch.org/get-started/locally/
